Transferring all used functions from use-app-url to urls.ts (#804)
Closes
https://github.com/SoftUni-Internal/exam-systems-issues/issues/877

**Summary of the changes made**:
1. Transferring all used functions from use-app-url to urls.ts
2. Removing the hook from App.tsx and deleting it
