+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/guidelines/alert/Alert.module.scss
@@ -1,7 +1,7 @@
 .success-alert {
-    background-color: (245, 78, 84)
+  background-color: (245, 78, 84);
 }
 
 .alert-button {
-    font-size: 18;
+  font-size: 18;
 }
