+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/guidelines/collapsible/Collapsible.module.scss
@@ -1,15 +1,14 @@
 .collapsibleContainer {
-  width: 100%;
   overflow: hidden;
   transition: max-height 0.55s ease-out;
+  width: 100%;
 
   &.visible {
-    max-height: max-content;
     margin: 5px 0;
+    max-height: max-content;
   }
 
   &.hidden {
     max-height: 0;
   }
 }
-
