+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/details/SubmissionDetails.tsx
@@ -8,13 +8,13 @@ import { DEFAULT_PROBLEM_RESULTS_TAKE_CONTESTS_PAGE } from '../../../common/cons
 import { contestParticipationType } from '../../../common/contest-helpers';
 import { useHashUrlParams } from '../../../hooks/common/use-hash-url-params';
 import { useSubmissionsDetails } from '../../../hooks/submissions/use-submissions-details';
-import { useAppUrls } from '../../../hooks/use-app-urls';
 import { useAuth } from '../../../hooks/use-auth';
 import { useContests } from '../../../hooks/use-contests';
 import { usePageTitles } from '../../../hooks/use-page-titles';
 import concatClassNames from '../../../utils/class-names';
 import { preciseFormatDate } from '../../../utils/dates';
 import { flexCenterObjectStyles } from '../../../utils/object-utils';
+import { getAdministrationRetestSubmissionInternalUrl, getParticipateInContestUrl } from '../../../utils/urls';
 import CodeEditor from '../../code-editor/CodeEditor';
 import AlertBox, { AlertBoxType } from '../../guidelines/alert-box/AlertBox';
 import { Button, ButtonSize, ButtonState, ButtonType, LinkButton, LinkButtonType } from '../../guidelines/buttons/Button';
@@ -50,7 +50,6 @@ const SubmissionDetails = () => {
     const { state: { user: { permissions: { canAccessAdministration } } } } = useAuth();
     const { actions: { loadContestByProblemId } } = useContests();
 
-    const { getAdministrationRetestSubmissionInternalUrl, getParticipateInContestUrl } = useAppUrls();
     const { state: { user } } = useAuth();
     const { state: { hashParam } } = useHashUrlParams();
     const navigate = useNavigate();
@@ -192,7 +191,7 @@ const SubmissionDetails = () => {
                 />
             );
         },
-        [ canAccessAdministration, getAdministrationRetestSubmissionInternalUrl ],
+        [ canAccessAdministration ],
     );
     const renderButtonsSection = useCallback(() => (
         <div className={styles.buttonsSection}>
@@ -314,7 +313,7 @@ const SubmissionDetails = () => {
             setCurrentSubmission(null);
             selectSubmissionById(null);
         },
-        [ currentSubmission, getParticipateInContestUrl, hashParam, navigate, selectSubmissionById, setCurrentSubmission ],
+        [ currentSubmission, hashParam, navigate, selectSubmissionById, setCurrentSubmission ],
     );
 
     const codeEditor = useCallback(
