+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/details/submission-result-details/SubmissionResultsDetails.module.scss
@@ -39,13 +39,13 @@
 .showDetailsButton {
   background-color: colors.$light-blue-color;
   border: 0;
+  border-radius: 6px;
   color: colors.$white-color;
   cursor: pointer;
   display: inline-block;
   margin: 15px 0;
   padding: 5px 10px;
   width: 10%;
-  border-radius: 6px;
 }
 
 .warningBlockWrapper {
