+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/submission-grid-row/SubmissionGridRow.module.scss
@@ -63,6 +63,26 @@
     font-size: variables.$f-size-16;
   }
 
+  .competeIcon {
+    margin: 0 10px;
+  }
+
+  .competeIconHoverText {
+    background-color: colors.$white-color;
+    border: 1px solid colors.$light-blue-color;
+    border-radius: 4px;
+    color: colors.$light-blue-color;
+    font-weight: 400;
+    opacity: 0;
+    padding: 5px 10px;
+    position: absolute;
+    text-align: center;
+    top: -40px;
+    transform: translateX(-50%);
+    transition: all 0.3s ease;
+    white-space: nowrap;
+  }
+
   .competeIconWrapper {
     position: relative;
 
@@ -72,25 +92,4 @@
       }
     }
   }
-
-  .competeIcon {
-    margin: 0 10px;
-  }
-
-  .competeIconHoverText {
-    opacity: 0;
-    position: absolute;
-    top: -40px;
-    transform: translateX(-50%);
-    background-color: colors.$white-color;
-    white-space: nowrap;
-    border: 1px solid colors.$light-blue-color;
-    border-radius: 4px;
-    color: colors.$light-blue-color;
-    font-weight: 400;
-    padding: 5px 10px;
-    text-align: center;
-    transition: all 0.3s ease;
-  }
 }
-
