+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/submissions-grid/SubmissionsGrid.module.scss
@@ -15,17 +15,18 @@
 
   button {
     @extend %font-family-montserrat;
-    border: none;
+
+    background-color: colors.$white-color;
+    border: 0;
     border-radius: 8px;
+    box-shadow: colors.$box-shadow-color;
+    color: colors.$light-blue-color;
     cursor: pointer;
-    margin: 30px 0 10px 0;
-    padding: 13px 24px;
     font-size: variables.$f-size-16;
     font-weight: 700;
-    color: colors.$light-blue-color;
-    background-color: colors.$white-color;
-    box-shadow: colors.$box-shadow-color;
-    transition: all .2s ease-in-out;
+    margin: 30px 0 10px;
+    padding: 13px 24px;
+    transition: all 0.2s ease-in-out;
 
     &:first-child {
       margin-right: 20px;
@@ -40,14 +41,15 @@
 
 .noSubmissionsFound {
   @extend %font-family-montserrat;
+
+  align-items: center;
   display: flex;
   justify-content: center;
-  align-items: center;
   margin-top: 20px;
 }
 
 .filtersContainer {
-  display: flex;
   align-items: center;
+  display: flex;
   justify-content: space-between;
 }
