+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/submissions-list/SubmissionsList.module.scss
@@ -3,8 +3,8 @@
 
 .submissionsList {
   align-items: start;
-  padding: 20px 0;
   margin-right: 20px;
+  padding: 20px 0;
 
   .submissionListItem {
     color: colors.$text-dark-gray-color;
